Add support for side-by-side static routes and dynamic routes
You can now define routes like so:

/a
/b
/c
/{else}

and also:

/{dynamic}
/{dynamic}/a
/{dynamic}/b

We also take care to support:

/x/y
/{else}

where "/x" will resolve to the "/{else}" case and
not get consumed as an impartial match by the "/x/y"
case.

Finally, it's worth calling attention to this ambiguous
case:

/a/{b}
/{a}/b
/{a}/{b}

Above, things will be matched in that order -- the
earlier a static route shows up in the match, the
higher its matching precedence. So "/a/b" will
always match "/a/{b}" and not "/{a}/b"